# Add flag for enabling finer-grained cuda graph capture