--- a/vllm/config.py
+++ b/vllm/config.py
@@ -3312,8 +3312,12 @@ class VllmConfig:
 
                 possible_sizes = [1, 2, 4] + [8 * i for i in range(1, 1025)]
                 if envs.VLLM_CAPTURE_ADDITIONAL_SMALL_GRAPHS:
-                    logger.info("Capturing additional cuda graphs. This will significantly increase the GPU memory utilization")
-                    possible_sizes = sorted(set(possible_sizes + list(range(1, 8)) + list(range(8, 16, 2)) + list(range(16, 32, 4))))
+                    logger.info(
+                        "Capturing additional cuda graphs. This will significantly increase the GPU memory utilization"
+                    )
+                    possible_sizes = sorted(
+                        set(possible_sizes + list(range(1, 8)) +
+                            list(range(8, 16, 2)) + list(range(16, 32, 4))))
                 # find the minimum size that is larger than max_num_seqs,
                 # which then becomes the max_batchsize_to_capture
                 larger_sizes = [