--- a/vllm/envs.py
+++ b/vllm/envs.py
@@ -577,11 +577,11 @@ environment_variables: Dict[str, Callable[[], Any]] = {
     # in addition to the default set of sizes. The sizes are specified
     # as a comma-separated sequence of integers.
     "VLLM_EXTRA_CUDA_GRAPH_SIZES":
-    lambda: list(
-        map(
-            int,
-            filter(None,
-                   os.getenv("VLLM_EXTRA_CUDA_GRAPH_SIZES", "").split(',')))),
+    lambda: [
+        int(s.strip())
+        for s in os.getenv("VLLM_EXTRA_CUDA_GRAPH_SIZES", "").split(',')
+        if s != ""
+    ]
 }
 
 # end-env-vars-definition