For certain batch sizes, the rounding effect of cuda graph padding can be severe. Consider for instance a batch of size 9, rounded up to the nearest multiple of 8 (16), can lead to a noticeable worsening in performance around batch sizes, especially at batch_size=5 (rounded up to 8) and batch_size=9 (rounded up to 16).

This PR adds an opt-in flag to capture additional CUDA graphs at small batch sizes to offset this problem.

I benchmarked using a deployment of DeepSeek-R1 on 8xH200, with `max-concurrent==9` and `num-requests==100`. 

Without this PR, I measured **TPOT 55ms, Output Token Throughput 154 tok/s**
With the flag enabled, I measured **TPOT 44ms, Output Token Throughput 192 tok/s**, for an improvement of **~25%**.