@simon-mo 2025-02-07
On the API front, I think a more general approach can be an override to pass in the required cuda graph sizes, e.g. 
```
VLLM_OVERRIDE_CUDA_GRAPH_SIZES = "1,2,3,4,5,6,7,8,9,10,11,16,24,32"
```

as something you can can parse by `[int(s.strip()) for s in inp.split(",")]`.