@mgoin 2025-02-07
Maybe `VLLM_EXTRA_CUDA_GRAPH_SIZES` or `ADDITIONAL` (too long) so it is merged with the existing set, rather than fully overriding