@benchislett 2025-02-10
Thank you @simon-mo and @mgoin for the feedback, I agree that the proposed interface is better and have updated my PR to implement this.