@youkaichao 2025-02-11
overriding cudagraph sizes is already supported, although it's not well documented 👀 

you can use `-O {"cudagraph_capture_sizes": [1, 2, 4, 8, 10, ...]}`